Adding cache functionality to auth extensions
