--- Snacks.Entity.Authorization/Extensions/TypeExtensions.cs
@@ -1,35 +1,0 @@
-﻿// https://stackoverflow.com/a/1075059
-using System;
-
-namespace Snacks.Entity.Authorization.Extensions
-{
-    static class TypeExtensions
-    {
-        public static bool IsAssignableToGenericType(this Type givenType, Type genericType)
-        {
-            var interfaceTypes = givenType.GetInterfaces();
-
-            foreach (var it in interfaceTypes)
-            {
-                if (it.IsGenericType && it.GetGenericTypeDefinition() == genericType)
-                {
-                    return true;
-                }
-            }
-
-            if (givenType.IsGenericType && givenType.GetGenericTypeDefinition() == genericType)
-            {
-                return true;
-            }
-
-            Type baseType = givenType.BaseType;
-
-            if (baseType == null)
-            {
-                return false;
-            }
-
-            return IsAssignableToGenericType(baseType, genericType);
-        }
-    }
-}
