+++ Snacks.Entity.Authorization/Snacks.Entity.Authorization.csproj
@@ -4,14 +4,14 @@
     <TargetFramework>netcoreapp3.1</TargetFramework>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
     <PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
-    <Version>2.0.0</Version>
+    <Version>2.0.1</Version>
     <Description>Authorization extensions for the Snacks.Entity.Core package.</Description>
     <Copyright>2020 - Evan Irla</Copyright>
     <PackageLicenseFile>LICENSE</PackageLicenseFile>
     <PackageProjectUrl>https://github.com/evanirla/snacks-entity-authorization</PackageProjectUrl>
     <RepositoryUrl>https://github.com/evanirla/snacks-entity-authorization</RepositoryUrl>
     <RepositoryType>Git</RepositoryType>
-    <PackageTags>entity-framework, authorization</PackageTags>
+    <PackageTags>ef-core, authorization</PackageTags>
   </PropertyGroup>
 
   <ItemGroup>
@@ -20,6 +20,8 @@
     <PackageReference Include="Microsoft.AspNetCore.Http.Features" Version="3.1.11" />
     <PackageReference Include="Microsoft.AspNetCore.Mvc.Core" Version="2.2.5" />
     <PackageReference Include="Microsoft.Extensions.Caching.Abstractions" Version="3.1.11" />
+    <PackageReference Include="Snacks.Entity.Caching" Version="2.0.1" />
+    <PackageReference Include="Snacks.Entity.Core" Version="2.0.1" />
   </ItemGroup>
 
   <ItemGroup>
@@ -27,12 +29,6 @@
       <Pack>True</Pack>
       <PackagePath></PackagePath>
     </None>
-  </ItemGroup>
-
-  <ItemGroup>
-    <Reference Include="Snacks.Entity.Core">
-      <HintPath>..\..\snacks-entity-core\Snacks.Entity.Core\bin\Debug\netcoreapp3.1\Snacks.Entity.Core.dll</HintPath>
-    </Reference>
   </ItemGroup>
 
 </Project>
